support "-zsh" as zsh
fixes #20

Separating our shell_types to it's own module to improve readability
of bloated main function.

Modifying test structure to match more modern patterns (tests directory
matching src structure)
