+++ src/commands/mod.rs
@@ -1,8 +1,6 @@
 mod help;
 mod init;
 mod script;
-#[cfg(test)]
-mod tests;
 pub use self::help::help;
 pub use self::init::init;
 pub use self::script::Script;
